apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.2.1
duckdb==0.10.0
redis==5.0.1
pyarrow==15.0.0
pandas==2.2.0
openai==1.10.0
sqlglot==20.9.0
pytest==8.0.0
httpx==0.26.0
python-multipart==0.0.9
pyyaml==6.0.1
fastapi-cache2[redis]==0.2.1
